+++ appCitas/app/Http/Controllers/HospitalController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class HospitalController extends Controller
+{
+    //
+}
