+++ appCitas/app/Http/Controllers/MedicoController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class MedicoController extends Controller
+{
+    //
+}
