+++ appCitas/app/Models/Paciente.php
@@ -17,4 +17,8 @@ class Paciente extends Model
         'fecha_nacimiento',
         'user_id',
     ];
+    public function user()
+    {
+        return $this->belongsTo(User::class);
+    }
 }
